Functioning refactor
