+++ identity_jose/src/jws/mod.rs
@@ -9,12 +9,11 @@
 mod algorithm;
 mod charset;
 mod decoder;
-mod decoder_config;
 mod encoder;
 mod format;
 mod header;
-mod jws_verifier;
 mod recipient;
+mod verification_plug_ins;
 
 pub use self::algorithm::*;
 pub use self::charset::*;
@@ -22,5 +21,5 @@ pub use self::decoder::*;
 pub use self::encoder::*;
 pub use self::format::*;
 pub use self::header::*;
-pub use self::jws_verifier::*;
 pub use self::recipient::*;
+pub use self::verification_plug_ins::*;
