+++ identity_jose/src/jws/verification_plug_ins/mod.rs
@@ -1,0 +1,6 @@
+// Copyright 2020-2023 IOTA Stiftung
+// SPDX-License-Identifier: Apache-2.0
+mod error;
+mod jws_verifier;
+pub use error::*;
+pub use jws_verifier::*;
